Update DetallePedidoDao.java
Cambio en el metodo:"public List<DetallePedido> traerPorPlato(int idPlato)"
El parametro paso de ser long a int

+++ src/dao/DetallePedidoDao.java
@@ -84,7 +84,7 @@ public class DetallePedidoDao {
 		return lista;
 	}
 	
-	public List<DetallePedido> traerPorPlato(long idPlato) throws HibernateException {
+	public List<DetallePedido> traerPorPlato(int idPlato) throws HibernateException {
 
 		List<DetallePedido> lista = null;
